requirements.txt
google-auth==2.27.0
google-auth-oauthlib==1.2.0
google-api-python-client==2.111.0
pandas==2.1.3
matplotlib==3.8.2
seaborn==0.13.1
plotly==5.18.0
dash==2.14.2
python-dotenv==1.0.0
streamlit==1.28.0
